Updating "Issues Addressed" Taxonomy
Content for the "Topics" questions. Using the Democracy Labs terms, with a few additional synonyms. Also using the term "Issues Addressed" rather than "Topics" as that is more specific and leads to the aggregations we're looking for with this taxonomy.

Added a folder for individual files, and removed the "Topics" file in the master branch.

This will serve as the baseline for reviewing and proposing changes to the taxonomy.